Desktop/telegram bot/requirements.txt
pip==25.1.1
setuptools>=65.5.0
wheel>=0.40.0
python-telegram-bot==20.3
Pillow==10.0.0